backend/core/pyproject.toml
[tool.poetry]
name = "market"
version = "0.1.0"
description = ""
authors = []

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test*.py"
pythonpath = ". src"

[tool.black]
line-length = 85
skip-string-normalization = true

[tool.ruff]
line-length = 100
src = ["src"]
indent-width = 4
target-version = "py311"
fix = true
exclude = [
    "migrations",
    "manage.py",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "__pychache__",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests"
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ARG002",
    "ANN",
    "D",
    "PLR0913",
    "RUF001",
    "G004",
    "PT009",
    "DJ001",
    "RUF012",
    "DTZ003",
    "PLR2004",
    "RET505",
    "RET503",
    "S113",
    "FBT001",
    "TRY400",
    "TRY201",
    "BLE001",
    "S101",
    "PT006",
    "TRY003",
    "EM101",
    "A002"
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
line-ending = "lf"
quote-style = "single"
docstring-code-format = true
skip-magic-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
case-sensitive = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.poetry.dependencies]
python = "~3.11"
django = "^4.1.6"
django-cachalot = "^2.5.3"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
redis = "^4.5.5"
django-redis = "^5.3.0"
django-cors-headers = "^4.2.0"
pre-commit = "^3.3.3"
requests = "^2.31.0"
pika = "^1.3.2"
dependency-injector = "^4.41.0"
yookassa = "^3.5.0"
sentry-sdk = "^2.37.1"
faststream = {extras = ["rabbit"], version = "^0.5.48"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"
pytest-django = "^4.11.1"
pytest = "^9.0.2"
